examples/terminal/requirements.txt
prettytable==3.6.0
requests==2.28.2
python-dotenv==1.0.0
boto3==1.26.85